v0.37.0 — production hardening: 25 unwrap fixes, structured logging, release profile, clippy-clean
Phase 1: Fix 25 unwrap/panic sites
- cluster.rs: 15 mutex .lock().unwrap() → .unwrap_or_else(|e| e.into_inner())
- feature_flags.rs: 8 mutex poison-recovery conversions
- entity_extract.rs: safe Option handling for parts.last()
- storage.rs: match on Option instead of clone().unwrap()

Phase 2: Structured logging
- Added log 0.4 + env_logger 0.11 dependencies
- env_logger::init() in main.rs; controlled via RUST_LOG env var

Phase 3: Release profile & lints
- [profile.release]: lto=true, codegen-units=1, strip=true (binary: 8.6 MB)
- [lints.rust]: unsafe_code=forbid
- [lints.clippy]: all=warn; resolved 220 warnings to 0
- Fixes across beacon, fingerprint, kernel_events, quantum, ws_stream, server

Phase 4: Version sync
- Cargo.toml, Helm Chart.yaml, README, STATUS, ROADMAP → 0.37.0

Phase 5: 9 new tests (929 lib total)
- Cluster: concurrent ops, commit-index, election check
- Feature flags: kill-switch, concurrent stress, unknown-flag safety
- Entity extract: empty input, domain edge cases, no-IP verification

58 files changed across all phases.

+++ src/notifications.rs
@@ -290,7 +290,7 @@ impl NotificationEngine {
                         last_err = Some(e);
                         if attempt + 1 < self.max_retries {
                             std::thread::sleep(std::time::Duration::from_millis(
-                                100 * 2u64.pow(attempt as u32),
+                                100 * 2u64.pow(attempt),
                             ));
                         }
                         continue;
